Add `now.env` in package.json as env source
This adds package.json as new source for environment variables, which
are merged with `-e` CLI arguments (CLI args take precedence).

With the ability to specify environment variables in package.json,
developers don't need to remember which `-e` arguments to add, and
can simply run `now`.

```json
"now": {
  "env": {
    "foo": "bar"
  }
}
```

The above is the same as running:
```
now -e foo=bar
```

Fixes #116