don't include the "main" file by default
It's not correct to forcefully include the file specified in "main"
at upload-time. Consider an `npm run build` script that ends up
creating the file that "main" specifies. It would not exist until
inside the deployment, after the build script is ran for that
deployment.

Fixes https://github.com/zeit/now-cli/issues/281.