secrets: treat any 4xx error code as "user error"
The Secrets endpoints used to inappropriately return a 400
error code when a secret name collision occurred (i.e. when
creating a secret with a name that already exists, or when
trying to rename a secret to a name that is already in use).
Now they return 409 Conflict which is more correct for those
scenarios.

So while we're at it, just treat all 4xx error codes as user errors.