Add url in the final message
Hi guys,
it would be wonderful if I can access the url when the deployment is complete. 
I don't like to scroll it up just to copy the url. 
I know it copies for me at beginning, but while it is deploying I am still working and in the vast majority of times I have another thing in my clipboard when the deployment end.