Remove `--base64` argument from `now secret add`
It's been broken on the server-side for quite a while now,
and on top of that it's never been clear as to *how* you
get the binary value back out from within the deployment.

For these reasons, let's get rid of the option altogether.

Users who need to encode binary data as the secret value
should encode to base64 manually. For example:

```
$ now secret add my-secret $(echo 'foo' | base64)
```