Add `now init` command
This command inspects a directory, and uses various heuristics
to determine the "project type" as well as generated an ideal
initial Dockerfile to use for the detected project type.

A `now.json` file is also created to set the proper "type"
(`docker` or `static`), as well as enable Cloud v2.

At the moment, only Node.js projects are supported, but the
`@zeit/init` module will be improved to support many project types,
including static site generators i.e. `next export`.