[now-static-build] Use `stdio: 'inherit'` for "dev" script child process
Since `@now/static-build` is no longer sniffing the stdio streams for
the bound port number in `now dev`, there's no need to have separate
stdio streams for the "dev" script. Instead, inherit stdio from the
parent process, which will allow for ANSI colors to be used when stdout
is a TTY in `now dev`.

Also simplifies the `checkForPort()` function and removes
`promise-timeout` dependency.