[now-cli] Purge require cache after packages install in `now dev`
This is really only relevant for the unit tests, where multiple
`DevServer` instances are create in the same process. But this avoids
the parent `now dev` process from having stale builder information
after a different version of the same builder is installed.

Fixes builder 2 vs. 3 discrepancies that are causing unit tests to fail.