[now-node] Write the dev mode tmp tsconfig.json file to $TMP
Instead of polluting the user's working directory, store the JSON file
in a temporary location.

Note: this code doesn't work as-is (when the `.json` file is in tmp,
then it doesn't have access to the projects `node_modules` and thus,
is missing typings. Discussed on a call with @leo and @styfle to place
the file in the project's `.now` directory, which is still TODO.