[now-cli] Support WebSocket upgrades in `now dev`
We were already using `http-proxy`'s WebSockets option, but we were missing a step! In addition to setting `{ws: true}`, you must also forward WebSocket upgrade events from the front-facing server to the proxy.

Previously, the proxy server and client would mutually agree to "upgrade" their connection (because we set `{ws: true}`)… but, when the client would immediately call back with an upgrade event, our front-facing server would drop it because there was no WebSocket upgrade handler.

In this change, we configure `this.server` to forward its `upgrade` events to the proxy server.

We also refactor `proxyPass` to use a new shared `this.proxy` instance, instead of creating a new proxy instance for every request.

This fixes #3451: CRA apps can now auto-refresh when the source code changes, and Firefox can now load the app without crashing on the 6th attempt.