refactor file reading & create failing testcase
This commit refactors out the file tree building from create-deployment.ts into its own functions and adds a simple unit test for each. The most important of the two is the test for readdirRelative because before v18 of the cli it didn't walk through node_modules but now it does which broke one of my deployments. The test for readdirRelative's should fail in this commit to prove the issue exists and then I will fix it in the next commit.

Address nits