[cli] Add `vc dev --update-runtimes` option
This is an experiment to see if the updating logic is causing the race
condition that sometimes causes Runtimes to fail with "module not found"
errors. My hypothesis is that a `require()` happens for a runtime right
when npm is overwriting the module with a new version, causing files to
be missing and lead to the error.