[cli] Default to `os.EOL` when creating `.gitignore` file
This fixes the scenario where 1) is Windows and 2) there is no
`.gitignore` file in the project, so `vc link` creates one. Before,
it was defaulting to `\n`. Updated to use `os.EOL` so that `\r\n` is
used on Windows.

Also removed `encoding` option to `writeFile()` since `utf8` is already
the default.

Related to #4965.