[cli] Ensure that the spinner is started when setting `text`
There was an issue with `output.spinner()` spinner if the old-style
interface was used (invoking the return value to stop the spinner),
which caused the internal spinner instance in `output` to be defined,
but in a "stopped" state.

Then later the `text` gets updated to a different value but the spinner
is still stopped and nothing gets rendered.

The fix is to ensure that `spinner.start()` is called when the `text`
property is updated.