[cli] Print login URL to terminal
In some cases (i.e. when SSH'd to a remote machine) the `open` command
will not work reliably. So we need to print the URL to the user as
a fallback for those cases when the web browser is not automatically
opened.

This also moves where `tokenName` is specified to be in the "verify"
endpoint, so that it does not need to be part of the URL that gets
printed to the user.