[cli] Allow dynamic label in `output.time()` function
Updates the `output.time()` function to accept a function as the "label"
parameter. When a function is passed, it will be invoked twice, once at
the beginning of the promise being executed, and a second time after the
promise has resolved. The second time, the return value of the promise
is passed to the label function, so that the resolved value may be used
to create the label.

Drops usage of Node's `console.time()` and `console.timeEnd()` since we
were using it in a hacky and unnecessary way.