[cli] Add request counter to `--debug` logs
Follow up to #6508. Adds a request "counter" that increments for every
API request, matching how `front` does it.

Also tweaks the `output.time()` output a bit by rendering the timestamp
in gray, and with `[]` brackets. If the time diff is < 1000ms then the
full milliseconds will be printed intstead of the seconds.