[build-utils] Add `launcher` config to Lambda class
The `launcher` property can be used to signal to the build system that
it needs to add additional files to the final Lambda before creating the
final zip file. It's a generic property so that different runtimes will
be able to specify specific configuration options that the build system
will utilize in order to add the necessary files.