[cli] Refactor vc build `installBuilders()` logic to not use a loop
This will prevent the possibility of an infinite loop occurring when
importing builders that throw a `MODULE_NOT_FOUND` error. Instead, we'll
just try to import the builders, then run yarn if necessary, and then
try to import the builders a second time when needed. If that fails then
we'll throw the original error.