[build-utils] Return `middlewareRoutes` from `detectBuilders()`
Instead of `@vercel/node` being responsible for creating the middleware
route, we can have it be part of the zero-config detection logic.

This will have the advantage of `vercel dev` being able to get that route
since it runs the `detectBuilders()` function, and also seems like
a cleaner separation of concerns since `@vercel/node` will not need be
responsible for generating any routes, so the `routes` property is
removed from the `BuildResultV3` type once again.