If the test is failing, just remove it :exclamation: 🧠
This reverts commit 8eb12fc0cebc2f981acff0c2f0fa4d10a6e0f3eb.

This test is failing because you need to have a Git provider Login Connection on your account in order to be able to connect a Git provider repo. There's no good way to do this within integration tests, so I'm removing the test for now and will instead figure out how to accept mock inputs in unit tests.