[node] Update to TypeScript v4.7.4
Note that this is a BREAKING CHANGE since it will change the default
version of TypeScript that is used for user Vercel deployments to this
version.

We should schedule a date for this change to happen, pre-announce that
the change in default version is coming, and publish a changelog once it
happens.