[cli] Replace "progress" module with custom implementation
The `progress` module is overkill for what we are using it for and it's
inconsistent with our styling for the rest of the deployment process.

So drop the npm module in favor of an in-house progress bar
implementation that can be used inside of the Ora `output.spinner()`
that we already use for the other steps of the deployment process.