[cli] Replace "0.0.0.0" with "localhost" in `vc dev`
Instead of replacing "127.0.0.1" in `vc dev`, replace "0.0.0.0" when
rendering the URL to access the `vc dev` server.

The thought here is that the IPv6 form is replacing the catch-all host
name (`::`), but the IPv4 version is only replacing the explicit
"127.0.0.1" version, so there's an inconsistency.

Considering the catch-all host names are not usually intended to be
routable anyways, we will fix the inconsistency by only replacing the
catch-all versions instead of localhost.

Also, since Node.js will default to binding to the catch-all address
when not explicitly specifying a host, this replacement will only happen
in that case. If the user is explicitly specifing a host to bind to
(i.e. `vc dev -l 127.0.0.1:8080` or `vc dev -l tcp://[::1]:8080`) then
it makes more sense to print explicitly what the user specified.