Update frameworks.ts for Next.js 13
When deploying a Next.js app with Vercel, the output seems to point to 'public' as the default directory instead of 'pages'. This results in a `404: NOT_FOUND` error when displaying the page. This is as confirmed by the placeholder value in ./packages/frameworks/src/frameworks.ts for Next.js which is set to 'Next.js default'