[remix] Allow `runtime` to be defined from a parent route
Make it so that the `export const config = { runtime }` value is
"inherited" from a parent route, if it is defined.

For example, "edge" can be defined in `app/root.tsx`, and so all routes
will use the Edge runtime by default, unless `"runtime": "nodejs"` is
used more specifically in a route deeper in the route hierarchy.