[remix] Simplify ESM detection of `remix.config.js`
In a Turborepo setup, there was this in the `remix.config.js` file:

```
  watchPaths: [require.resolve("ui")],
```

Since we attempt to `require()` the `remix.config.js` file in order to determine whether or not it is using ESM syntax, this would fail since the require happens before the Build Command is executed.

To fix, treat any non `ERR_REQUIRE_ESM` error as CJS, instead of
re-throwing the error. If there really is an issue with importing the
config, then that'll happen right after when `remix build` is doing it.