[remix] `cd` to the project directory during `readConfig()`
Sometimes the `cwd` needs to be properly set for some Remix
configurations. This is problematic in monorepos because the default
`cwd` is the root of the monorepo, and not the project directory.

So `cd` to the project directory when reading the Remix config file to
account for that.

Fixes https://github.com/kiliman/remix-flat-routes/issues/42.
Fixes https://github.com/vercel/vercel/discussions/9478.