[tests] Properly test for multiple headers with same name in probes
Updates the `responseHeaders` probe checks to properly test for multiple
headers with the same name.

Previously the probes were using `headers.get()` which concats multiple
headers into a single string, which results in the check not really
checking if there are in fact multiple headers with the same name.
Using `headers.raw()` allows us to properly test for that.

A couple of Python tests that were already checking for multiple
`set-cookie` headers needed to be updated to match the full value, since
the check now properly validates the full string match of each header
(before it was basically just doing a `string.includes()` check).

This is a precursor for https://github.com/vercel/vercel/pull/9533.