[remix] Inject Edge entrypoint into `@remix-run/vercel` during Edge Function creation
There's a PR opened on Remix to add an Edge compatible entrypoint to @remix-run/vercel. This is only really needed specifically for projects that have a custom server.js entrypoint file, which is not strictly necessary anymore, but there are some edge cases where a project might still want to have one. So this PR is necessary for those kinds of projects to be able to use Edge runtime.

Even when they merge the PR, it would be good to leave this injection code around for some time to continue support for older @remix-run/vercel versions in existing projects.